refactor(ui): extract VST3 / export / pattern / diagnostics + Phase 0/7 handler domains from main.js (S5 Stage 5)
Final handler-domain stage of the ui/main.js decomposition. Move the last
extractable IPC handler domains verbatim out of ui/main.js into
ui/electron-main/. The 122-channel commit artifact is unchanged; channel
names, behavior and defaults are a pure move.

- phase0-compat.js — the flat legacy transport/frame/sync channels
  (xleth:play/stop/pause/trigger/transportState/currentFrame/frameRGBA/
  syncStats/setVideoResolution + proxy:getStatus) that predate the Phase-1
  domain scheme. Still load-bearing: preload.js routes BOTH its legacy flat
  API AND its namespaced transport/video/sync/proxy/sampler wrappers through
  these exact names, so they cannot be retired until preload stops invoking
  them (noted in the module header).
- quick-launchers.js — launcher .exe/.png pickers + detached spawn.
- preview-visibility.js — the Phase 7 xleth:preview:setEnabled toggle.
- patterns.js — all timeline_* region/syllable/pattern/pattern-block/note
  pass-throughs. The region/syllable group had been orphaned under a Stage-3
  clip-defaults breadcrumb; it shares the timeline_* domain, so it lives here.
- vst3.js — plugin scan/progress, editor-window open/close/query, missing/
  crashed-plugin recovery, and the add-search-path directory picker.
- export.js — audio + video export, both progress-poll intervals, the
  hardware-encoder + GPU queries, and isExportBusy() (consumed by the autosave
  gate; main.js's line-71 wiring now sources from it). The video handlers,
  physically split from their poll helper in main.js by the WORLD-poll
  section, are reunited here.
- diagnostics.js — the Visual Preview Diagnostic Log + six-stage pixel-content
  verification helpers and the xleth:diag:exportVisualPreviewLog handler.

Handlers that wrap safeHandler register inside init({ safeHandler }); dialog /
window-owning ones also take getWin. callWorker comes from the Stage 1
worker.js; diagnostics.js pulls loadSettings from settings.js and userDataPath
from runtimePaths. graphHandler, the adaptive WORLD-processing poll, the shm
video frame output, xleth:video:computeDurationSeconds and xleth:importVideo
stay inline in main.js (the WORLD poll + shm frame output are permanent per
AUDIT.md §4.3). main.js 2736 -> 1794 lines.

Verified: npm run build && npm run test:smoke -> 3/3 against both the working
tree and this exact committed main.js; vitest: 2478 passed, 2 failed = the
pre-existing local-only mixer FxBadge failures, identical to the Stage 1-4
baseline. IPC channel set verified identical (122) between HEAD main.js and
the commit artifact.
